requirement.txt
asgiref @ file:///home/conda/feedstock_root/build_artifacts/asgiref_1605523041333/work
certifi==2020.12.5
chardet==3.0.4
Django @ file:///home/conda/feedstock_root/build_artifacts/django_1609752033316/work
django-countries==6.1.3
django-filter==2.4.0
django-rest-passwordreset==1.1.0
djangorestframework @ file:///home/conda/feedstock_root/build_artifacts/djangorestframework_1609945773755/work
idna==2.10
mysqlclient==2.0.1
Pillow==8.0.1
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1608904108784/work
requests==2.25.0
sqlparse @ file:///home/conda/feedstock_root/build_artifacts/sqlparse_1602142927465/work
stripe==2.55.0
urllib3==1.26.2
wincertstore==0.2
